Remove track events from getWordPressInstallationState() method
